# pos-models/target/generated-sources/openapi/src/main/java/com/micropos/api/OrderApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (5.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.micropos.api;

import com.micropos.dto.ErrorDto;
import com.micropos.dto.ItemDto;
import java.util.List;
import com.micropos.dto.OrderDto;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-12-10T18:32:44.530194200+08:00[Asia/Shanghai]")
@Validated
@Tag(name = "order", description = "the order API")
public interface OrderApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /order : Add a new order
     *
     * @param itemDto  (optional)
     * @return Successful operation (status code 200)
     */
    @Operation(
        operationId = "createOrder",
        summary = "Add a new order",
        tags = { "cart" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  OrderDto.class)))
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/order",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<OrderDto> createOrder(
        @Parameter(name = "ItemDto", description = "", schema = @Schema(description = "")) @Valid @RequestBody(required = false) List<ItemDto> itemDto
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : \"id\", \"items\" : [ { \"quantity\" : 0.8008281904610115, \"productId\" : \"productId\" }, { \"quantity\" : 0.8008281904610115, \"productId\" : \"productId\" } ], \"status\" : \"status\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /order : All orders
     *
     * @return Successful operation (status code 200)
     *         or Cart not found (status code 404)
     */
    @Operation(
        operationId = "getAllOrders",
        summary = "All orders",
        tags = { "cart" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  OrderDto.class))),
            @ApiResponse(responseCode = "404", description = "Cart not found", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  ErrorDto.class)))
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/order",
        produces = { "application/json" }
    )
    default ResponseEntity<List<OrderDto>> getAllOrders(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : \"id\", \"items\" : [ { \"quantity\" : 0.8008281904610115, \"productId\" : \"productId\" }, { \"quantity\" : 0.8008281904610115, \"productId\" : \"productId\" } ], \"status\" : \"status\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /order/{orderId} : Get order information
     *
     * @param orderId The id of the order (required)
     * @return Successful operation (status code 200)
     *         or Order not found (status code 404)
     */
    @Operation(
        operationId = "getOrderById",
        summary = "Get order information",
        tags = { "cart" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  OrderDto.class))),
            @ApiResponse(responseCode = "404", description = "Order not found", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  ErrorDto.class)))
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/order/{orderId}",
        produces = { "application/json" }
    )
    default ResponseEntity<OrderDto> getOrderById(
        @Parameter(name = "orderId", description = "The id of the order", required = true, schema = @Schema(description = "")) @PathVariable("orderId") String orderId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : \"id\", \"items\" : [ { \"quantity\" : 0.8008281904610115, \"productId\" : \"productId\" }, { \"quantity\" : 0.8008281904610115, \"productId\" : \"productId\" } ], \"status\" : \"status\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /order/{orderId} : Remove order
     *
     * @param orderId Order ID (required)
     * @return Successful operation (status code 200)
     *         or Order not found (status code 404)
     */
    @Operation(
        operationId = "removeOrderById",
        summary = "Remove order",
        tags = { "cart" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  OrderDto.class))),
            @ApiResponse(responseCode = "404", description = "Order not found", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  ErrorDto.class)))
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/order/{orderId}",
        produces = { "application/json" }
    )
    default ResponseEntity<OrderDto> removeOrderById(
        @Parameter(name = "orderId", description = "Order ID", required = true, schema = @Schema(description = "")) @PathVariable("orderId") String orderId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : \"id\", \"items\" : [ { \"quantity\" : 0.8008281904610115, \"productId\" : \"productId\" }, { \"quantity\" : 0.8008281904610115, \"productId\" : \"productId\" } ], \"status\" : \"status\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /order/{orderId} : Update order status
     *
     * @param orderId Order ID (required)
     * @param status status (required)
     * @return Successful operation (status code 200)
     *         or Order not found (status code 404)
     */
    @Operation(
        operationId = "updateOrderStatusById",
        summary = "Update order status",
        tags = { "cart" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  OrderDto.class))),
            @ApiResponse(responseCode = "404", description = "Order not found", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  ErrorDto.class)))
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/order/{orderId}",
        produces = { "application/json" }
    )
    default ResponseEntity<OrderDto> updateOrderStatusById(
        @Parameter(name = "orderId", description = "Order ID", required = true, schema = @Schema(description = "")) @PathVariable("orderId") String orderId,
        @NotNull @Parameter(name = "status", description = "status", required = true, schema = @Schema(description = "")) @Valid @RequestParam(value = "status", required = true) String status
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : \"id\", \"items\" : [ { \"quantity\" : 0.8008281904610115, \"productId\" : \"productId\" }, { \"quantity\" : 0.8008281904610115, \"productId\" : \"productId\" } ], \"status\" : \"status\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}